% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildDesign}
\alias{buildDesign}
\title{Build the factorial design}
\usage{
buildDesign(shift_proportions, shift_magnitudes, shift_types, parameters_types,
  test_lengths)
}
\arguments{
\item{s_prop}{(numeric vector) Proportions of response shift.}

\item{s_magn}{(numeric vector) Magnitudes of response shift.}

\item{s_type}{(numeric vector) Types of response shift (i.e., towards slope, threshold, or both).}

\item{p_type}{(numeric vector) Types of item parameters (i.e., homogeneous or heterogeneous).}

\item{t_leng}{(numeric vector) Lengths of the test (i.e., number of items).}
}
\description{
Build the factorial design based on the vectors of independent variables.
}
