% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{effectSizeCohen}
\alias{effectSizeCohen}
\title{Compute Cohen's effect size}
\usage{
effectSizeCohen(firstVector, secondVector)
}
\arguments{
\item{firstVector}{(numeric vector) The first vector used in the effect size computations.}

\item{secondVector}{(numeric vector) The second vector used in the effect size computations.}
}
\description{
Compute the \code{d} effect size measure for two numeric vectors.
}
