% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParametersEstimation.R
\name{estimateParametersGrm}
\alias{estimateParametersGrm}
\title{Estimate the item paramters}
\usage{
estimateParametersGrm(data)
}
\arguments{
\item{data}{(numeric matrix) The data to fit the GRM on.}
}
\description{
Fit the GRM using \code{mirt} behind the scenes and 
return the simplified parameters. The output will 
hold the slope as the first column and from the
second the item steps.
}
