% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{FypSim}
\alias{FypSim}
\alias{FypSim-package}
\title{FypSim: Simulations for the First Year Paper}
\description{
FypSim provides an easy to use set of functions that allow you to
replicate the simulations conducted for the First Year Paper.
}
