% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCleaning.R
\name{glueDesignPieces}
\alias{glueDesignPieces}
\title{Function that allows to glue various design pieces together generated by \code{runCellRange}.
The pieces linked together and then averaged across the number of design replications.
Can also be used with just one design to obtain the average for the indicators.}
\usage{
glueDesignPieces(design, ..., saveTo = NULL)
}
\arguments{
\item{design}{(numeric matrix) The configuration of the factorial design that was used. The pieces that
will be linked together represent selections from this design matrix.}

\item{...}{(list of raw results) Contains the lists of the raw results as they are returned by \code{runCellRange} or
any other related function.}

\item{saveTo}{(string) Specifies the location where the averaged matrices of indicators are saved. If \code{NULL} then data
is not saved. Defaults to \code{NULL}.}
}
\value{
Returns a list of matrices representing the averaged indicators across the number of design iterations.
First index in the list holds the matrix of detection rates, second the matrix of means, and third the matrix
of standard deviations.
}
\description{
Function that allows to glue various design pieces together generated by \code{runCellRange}.
The pieces linked together and then averaged across the number of design replications.
Can also be used with just one design to obtain the average for the indicators.
}
