% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure.R
\name{initializeDesign}
\alias{initializeDesign}
\title{Internal function that initalizes the factorial design from the default values.}
\usage{
initializeDesign()
}
\description{
Internal function that initalizes the factorial design from the default values.
}
