% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runCell}
\alias{runCell}
\title{Apply procedure for a single cell}
\usage{
runCell(population_theta, shift_proportion, shift_magnitude, shift_type,
  parameters_type, test_length)
}
\arguments{
\item{population_theta}{(double) Vector of latent abilities for the population.}

\item{shift_proportion}{(double) Proportion of response shift.}

\item{shift_magnitude}{(double) Magnitude of response shift.}

\item{shift_type}{(int) Type of response shift (i.e., towards slope, threshold, or both).}

\item{parameters_type}{(int) Type of item parameters (i.e., homogeneous or heterogeneous).}

\item{test_length}{(int) Lengths of the test (i.e., number of items).}
}
\description{
Perform the calibration and estimation for a single cell. Subsequently, we will use this
when we loop through the entire factorial design and for each cell we instantiate the
SimulateCell class. We have to also keep an eye on the fact that the entire design
in itself will also need to be replicated \code{n} times.
}
