% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure.R
\name{runCellRange}
\alias{runCellRange}
\title{Functions that allow to devide the simulations across multiple PCs.
It stores both the data for the replications, and the designs.
Storing the designs might help us spot possible mistakes.}
\usage{
runCellRange(start, end, directory = ".", replications = 100)
}
\description{
Functions that allow to devide the simulations across multiple PCs.
It stores both the data for the replications, and the designs.
Storing the designs might help us spot possible mistakes.
}
