% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runSelectedCells}
\alias{runSelectedCells}
\title{Applies procedure for a selected number of cells}
\usage{
runSelectedCells(selected_cells)
}
\arguments{
\item{selected_cells}{(numeric matrix) The selected cell configurations to be ran, where
the columns respect the order indicated above.}
}
\description{
Performs the calibration and estimation for a selected number of cells.
The cells are passed as a numeric matrix where the columns are in the
following order: shift proportion, shift magnitude, shift type, the
type of the item parameters and the test length. Note, the cells
are ran only once. For a design replication we need loop over
this function. Thus, we'll have a separate wrapper for that.
As a return type, we have the following:
Rcpp list containing three numeric matrices: detection rate,
mean, and standard deviation. Each of these matrices are
structure the following way: the columns indicate the
theta level and row numbers points to the same row
from withing the selected cells matrix. Thus,
values in row \code{n} in any of these
matrices represents the value for
configuration at row \code{n}
in the selected cells.
}
