% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runSelectedCellsWithReplication}
\alias{runSelectedCellsWithReplication}
\title{Applies and replicates the procedure for a selected number of cells}
\usage{
runSelectedCellsWithReplication(selected_cells, design_replications)
}
\arguments{
\item{selected_cells}{(numeric matrix) The selected cell configurations to be ran, where
the columns respect the order indicated above.}

\item{design_replications}{(int) The number of times the selected cells will be replicated.}
}
\description{
Very similar with the runSelectedCells() function, the only difference
is that it also replicate the selected cells \code{n} times.
}
