% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildDesign}
\alias{buildDesign}
\title{Builds the factorial design}
\usage{
buildDesign(shift_proportions, shift_magnitudes, shift_types, parameters_types,
  test_lengths)
}
\arguments{
\item{shift_proportions}{(numeric vector) Proportions of response shift.}

\item{shift_magnitudes}{(numeric vector) Magnitudes of response shift.}

\item{shift_types}{(numeric vector) Types of response shift (i.e., towards slope, threshold, or both).}

\item{parameters_types}{(numeric vector) Types of item parameters (i.e., homogeneous or heterogeneous).}

\item{test_lengths}{(numeric vector) Lengths of the test (i.e., number of items).}
}
\description{
Builds the factorial design based on the vectors of independent variables.
}
